+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.15.3] - 2026-07-23
+
 ### Added
 - Keyboard shortcuts Ctrl/Cmd + = and Ctrl/Cmd + - step the UI scale up and down.